+++ ash/src/extensions/amd/mod.rs
@@ -1,5 +1,2 @@
-pub use self::buffer_marker::BufferMarker;
-pub use self::shader_info::{ShaderInfo, ShaderInfoResult};
-
-mod buffer_marker;
-mod shader_info;
+pub mod buffer_marker;
+pub mod shader_info;